Kombucha_Control/requirements.txt
mysql-connector
Adafruit_DHT
RPi.GPIO